make compile

+++ sqlgraph-mariadb/src/test/java/org/umlg/sqlgraph/test/tp3/SqlGraphProvider.java
@@ -3,6 +3,7 @@ package org.umlg.sqlgraph.test.tp3;
 import com.tinkerpop.gremlin.AbstractGraphProvider;
 import com.tinkerpop.gremlin.structure.Graph;
 import org.apache.commons.configuration.Configuration;
+import org.umlg.sqlgraph.sql.dialect.SqlDialect;
 import org.umlg.sqlgraph.structure.SqlGraph;
 import org.umlg.sqlgraph.structure.SqlGraphDataSource;
 
@@ -21,7 +22,7 @@ public class SqlGraphProvider extends AbstractGraphProvider {
     public Map<String, Object> getBaseConfiguration(final String graphName) {
         return new HashMap<String, Object>() {{
             put("gremlin.graph", SqlGraph.class.getName());
-            put("jdbc.driver", "org.mariadb.jdbc.Driver");
+            put("sql.dialect", "org.umlg.sqlgraph.sql.dialect.MariaDBDialect");
             put("jdbc.url", "jdbc:mysql://localhost:3306/" + graphName);
             put("jdbc.username", "test");
             put("jdbc.password", "password");
@@ -35,18 +36,16 @@ public class SqlGraphProvider extends AbstractGraphProvider {
                 g.tx().rollback();
             g.close();
         }
-        SqlGraphDialect sqlGraphDialect = null;
+        SqlDialect sqlDialect;
         try {
-            Class.forName(SqlGraphDialect.MARIADBDB.getJdbcDriver());
-            sqlGraphDialect = SqlGraphDialect.MARIADBDB;
-        } catch (ClassNotFoundException e) {
-        }
-        if (sqlGraphDialect == null) {
-            throw new IllegalStateException("Maria db driver " + SqlGraphDialect.MARIADBDB.getJdbcDriver() + " must be on the classpath!");
+            Class<?> sqlDialectClass = Class.forName(configuration.getString("sql.dialect"));
+            sqlDialect = (SqlDialect)sqlDialectClass.newInstance();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
         }
         try {
             SqlGraphDataSource.INSTANCE.setupDataSource(
-                    sqlGraphDialect.getJdbcDriver(),
+                    sqlDialect.getJdbcDriver(),
                     configuration.getString("jdbc.url"),
                     configuration.getString("jdbc.username"),
                     configuration.getString("jdbc.password"));
@@ -55,18 +54,22 @@ public class SqlGraphProvider extends AbstractGraphProvider {
         }
         try (Connection conn = SqlGraphDataSource.INSTANCE.get(configuration.getString("jdbc.url")).getConnection()) {
             DatabaseMetaData metadata = conn.getMetaData();
-            String jdbcUrl = configuration.getString("jdbc.url");
-            String catalog = jdbcUrl.substring(jdbcUrl.lastIndexOf("/") + 1);
-            String schemaPattern = null;
-            String tableNamePattern = "%";
-            String[] types = {"TABLE"};
-            ResultSet result = metadata.getTables(catalog, schemaPattern, tableNamePattern, types);
-            while (result.next()) {
-                StringBuilder sql = new StringBuilder("DROP TABLE ");
-                sql.append(sqlGraphDialect.getSqlDialect().maybeWrapInQoutes(result.getString(3)));
-                sql.append(" CASCADE;");
-                try (PreparedStatement preparedStatement = conn.prepareStatement(sql.toString())) {
-                    preparedStatement.executeUpdate();
+            if (sqlDialect.supportsCascade()) {
+                String catalog = "sqlgraphdb";
+                String schemaPattern = null;
+                String tableNamePattern = "%";
+                String[] types = {"TABLE"};
+                ResultSet result = metadata.getTables(catalog, schemaPattern, tableNamePattern, types);
+                while (result.next()) {
+                    StringBuilder sql = new StringBuilder("DROP TABLE ");
+                    sql.append(sqlDialect.maybeWrapInQoutes(result.getString(3)));
+                    sql.append(" CASCADE");
+                    if (sqlDialect.needsSemicolon()) {
+                        sql.append(";");
+                    }
+                    try (PreparedStatement preparedStatement = conn.prepareStatement(sql.toString())) {
+                        preparedStatement.executeUpdate();
+                    }
                 }
             }
         } catch (SQLException e) {
